fix: move common config to file
